Base layout ready for the project.
The client and server folders are ready to be worked on. The server folder is a node.js server that will be used to serve the client folder. The client folder is a react app that will be used to serve the front end of the application.
