+++ client/src/App.js
@@ -1,0 +1,11 @@
+import "./App.css";
+
+function App() {
+	return (
+		<div>
+			<h1>Code Dash</h1>
+		</div>
+	);
+}
+
+export default App;
